Require Sphinx >= 1.7 and update to new interface

+++ dev/tools/docs/build_html_brian2.py
@@ -1,9 +1,12 @@
 import os
 import shutil
 import sphinx
+if not sphinx.version_info >= (1, 7):
+    raise ImportError('Need sphinx version 1.7')
+from sphinx.cmd.build import main as sphinx_main
 import sys
 
 os.chdir('../../../docs_sphinx')
 if os.path.exists('../docs'):
     shutil.rmtree('../docs')
-sys.exit(sphinx.main(['sphinx-build', '-b', 'html', '.', '../docs']))
+sys.exit(sphinx_main(['-b', 'html', '.', '../docs']))
